Module with various partical radiation pyrometers types and several function like pyrometer's emissivity adjusting to make the "measured" temperature fit the real temperture

RadiationPyrometers.DefaultPyrometersTypes — Constant

Default pyrometers types

Base.isless — Method

Base.isless(p1::Pyrometer,p2::Pyrometer)

Vector of Pyrometer objects can be sorted using isless

RadiationPyrometers.fit_ϵ! — Method

fit_ϵ!(p::Pyrometer,Tmeasured::Float64,Treal::Float64)

Optimizes the emissivity of the pyrometer to make measured by the pyrometer temperature fit
fit the real temperature

Input:
p - pyrometer object
Treal - real temperature of the surface, Kelvins
Tmeasured - temperature measured by the pyrometer, Kelvins

RadiationPyrometers.fit_ϵ! — Method

fit_ϵ!(p::Vector{Pyrometer},Treal::Float64,Tmeasured::Vector{Float64})

Fits the emissivity of pyrometers to make measured temperature Tmeasured fit fit the real temperature Treal

Input: p - pyrometer objects vector , [Nx0] Treal - real temperature of the surface, Kelvins Tmeasured - temperatures measured by the pyrometers, in Kelvins, [Nx0]

RadiationPyrometers.fit_ϵ_wavelength! — Method

fit_ϵ_wavelength!(p::Pyrometer,Tmeasured::Float64,Treal::Float64)

Fits emissivity and returns it as a vector of the same size as pyrometer's wavelength region
Some pyrometers has 2-wavelength, other work on a single wavelength, for two-wavelength pyrometers
e_out will be a two-element vector
Input:
    p - pyrometer object
    Tmeasured - temperature measured by the pyrometer, Kelvins
    Treal - real temeprature of the surface, Kelvins

RadiationPyrometers.fit_ϵ_wavelength! — Method

fit_ϵ_wavelength!(p::Vector{Pyrometer},Treal::Float64,Tmeasured::Vector{Float64})

The same as fit_ϵ! except that it returns the vector of fitted emissivities of the same length to the total number of wavelength in all pyrometers in vaector p, e.g. if p[i] is the narrow-band pyrometer

RadiationPyrometers.full_wavelength_range — Method

full_wavelength_range()

Creates the wavelengths vector covered by all pyrometers in vector p

RadiationPyrometers.full_wavelength_range — Method

full_wavelength_range()

Creates the wavelengths vector covered by default pyrometers see DefaultPyrometersTypes

RadiationPyrometers.is_fixed_wavelength — Method

is_fixed_wavelength(::Pyrometer{N}) where N

True if Pyrometer is single wavelength

RadiationPyrometers.is_narrow_band — Method

is_narrow_band(p::Pyrometer)

True if pyrometer p is a narrow-band pyrometer (worsk on a fixed wavelengh region)

RadiationPyrometers.measure — Method

measure(p::Pyrometer,i::Float64)

Calculates the "measured" temperature from "mesaured" intensity by fitting the Planck function. The intensity units should be consistent with PlanckFunctions.ibb(λ,T) function for single wavelength pyrometer and with PlanckFunctions.bandpower(λ,T), it should be in [W/m²⋅sr⋅μm] Input: p - pyrometer object i - measured intensity in [W/(m²⋅sr⋅μm)] or integral (over wavelength spectral intensity) in [W/(m²⋅sr)] returns the temperature "measured" by this pyrometer (optional) Tstarting - starting temperature value

RadiationPyrometers.produce_pyrometers — Method

produce_pyrometers()

Creates the vector of all default pyrometers

RadiationPyrometers.set_emissivity — Method

set_emissivity(p::Pyrometer,em_value::Float64)

Setter for spectral emissivity

RadiationPyrometers.switch_the_type — Method

switch_the_type(λ::Float64)

Returns the type (which can be used as a key of DefaultPyrometersTypes dict) depending on the wavelengh
Input:
    λ - wavelength in μm

RadiationPyrometers.wavelengths_number — Method

wavelengths_number(p::Vector{Pyrometer})

Returns the total number of wavelength for the vector of pyrometers

RadiationPyrometers.wavelengths_number — Method

wavelength_number()

Returns the length of wavelengths vector covered by the DefaultPyrometersTypes (all default pyrometers wavelengh region)

RadiationPyrometers.wlength — Method

wlength(::Pyrometer{N}) where N

Returns the number of wavelengths